+++ castor.php
@@ -1,5 +1,7 @@
 <?php
 
+\define('CASTOR_USE_CHDIR', true);
+
 use Castor\Attribute\AsTask;
 
 use function Castor\context;